mr: Test the concept of demoting pages to WriteProtect
The idea here is to test the idea of demoting specific pages to
WriteProtect, by modifying their pte entry, to test if the page is
still in use.

A silly ioctl modifies the page back to WriteProtect, and the next
write to the page does cause another page fault to promote the page
back to Writeable.